fix(types): add MyPy type annotations across codebase
Adds missing return type annotations to all functions and methods.

Fixes 37 MyPy errors across source and test files.

Includes new testing documentation for CI workflows.

+++ src/sleepstack/mix_binaural_with_ambience.py
@@ -41,7 +41,7 @@ from typing import Tuple
 
 
 def db_to_gain(db: float) -> float:
-    return 10.0 ** (db / 20.0)
+    return float(10.0 ** (db / 20.0))
 
 
 def read_wav(path: str) -> Tuple[np.ndarray, int, int]:
